[generate_opcheck_tests] tests should ignore meta/FakeTensors
These tests generally don't work on meta tensors because they need to
compare the data of the Tensors. For example, SchemaCheckMode errors out
if any inputs are meta or Fake because it needs to check their storages
to see if any mutation occurred and those do not have storages.

Test Plan:
- new tests

ghstack-source-id: e7ead392dc40fe5165f3ebbe4a1c04c10155f5ae
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109641